+++ test/crypto_test.dart
@@ -12,5 +12,12 @@ void main() {
       expect(sha1Hash('test'), 'a94a8fe5ccb19ba61c4c0873d391e987982fbbd3');
       expect(sha1Hash('qsu-md5'), 'e5c5dc3b2be3542475671d460f906c3b176bb5bf');
     });
+
+    test('sha256Hash', () {
+      expect(sha256Hash('test'),
+          '9f86d081884c7d659a2feaa0c55ad015a3bf4f1b2b0b822cd15d6c15b0f00a08');
+      expect(sha256Hash('qsu-md5'),
+          '8c4cfec3ec79dc572958ea7f0e3cfd24b90d174969df9a4773b37b68498871ed');
+    });
   });
 }